resinOS-flash194: Update partition sizes, add padding
Padding needs to be added so that the xavier device
specific space is a multiple of 4096kb, and the
resin-boot partition is written at the exact next
available byte in the resulting image,

This way, the resulting image is an exact copy of what
is written to the device.

This change is done to align with jetson-flash
utility.

Signed-off-by: Alexandru Costache <alexandru@balena.io>