tegra194-flash-dry: Create boot0 blob for Xavier
Starting with L4T 32.4.2 Xavier also needs an
updated boot0 bundle for the mmcblk0boot0
partition, otherwise it will fail to boot
after the hostOs update. The resulting binary
currently does not need to be patched, as
in the case of TX2, but this may be necessary
in future L4Ts.

Signed-off-by: Alexandru Costche <alexandru@balena.io>