linux-firmware: Bump to latest upstream head revision
meta-balena PR 2240 is not ready yet as it leaves some boards
without rootfs space. We need to do one last TX2 release on 32.4.4
to include the latest supervisor version and then merge and
release the L4T 32.5.1 update.

This commit should be removed after the TX2 release,
before the linux-firmware update in meta-balena is merged.

During testing we observed that the ath10k and qca firmwares
that we bring in are older from what is in upstream master,
let's update to the latest version here until we update all
firmwares in meta-balena.

Signed-off-by: Alexandru Costache <alexandru@balena.io>