jetson-tx2-4-gb: Add TX2 4GB yocto machine
Layer meta-tegra already contains a jetson-tx2-4gb
yocto machine so we need to add another one with a different
name, which inherits the plain tx2 so that we don't need
to duplicate all the tx2 code. BBMasking the upstream machine
doesn't work with this yocto verison.

Deivce maintained by Andrew Smith

Signed-off-by: Alexandru Costache <alexandru@balena.io>