balena-image: Install tegra-boot-tools for Xavier based boards
The Xavier boards do not come with u-boot support. nvbootctrl
did not work for selecting slots, so let's use OE4T tegra-boot-tools
for selecting which slot - _a or _b to boot from. This package
will allow us to write the dtb and kernel for the new OS
to a single slot, and then mark it for booting.

Signed-off-by: Alexandru Costache <alexandru@balena.io>