recipes-bsp/tools: Add script for marking successful boot in rollback drop-in
Use the rollback-altboot to mark the current tegra
boot slot as succesfull, so that the next boots
will use this slot. If the service doesn't
get to run due to a kernel panic or similar boot
issue, the tegra bootloaders will switch to
booting the alternate slot after 3 unsuccessful
boots.

Signed-off-by: Alexandru Costache <alexandru@balena.io>