tegra-boot-tools: Enable slot redundancy, mark successful boots on Xavier AGX and Xavier NX
The Xavier line of boards don't have u-boot support
and we use the command line in the dtb to switch
between the two root filesystems. Let's use
A/B slot switch to enable altbooting on these
devices.

Once redundancy is enabled, each successful boot
needs to be marked and we do this in a service file,
otherwise the lowlevel tegra bootloaders will fallback
to loading the previous slot (old dtb and old cmdline)
after 3 boot sequences that have not been marked.

Signed-off-by: Alexandru Costache <alexandru@balena.io>