hostapp-update-hooks: Write new bootblobs to next active slot only
We'll use tegra-boot-tools to mark the next
active slot on the Xavier and write the new signed
bootblobs, which include the dtb that contains the
cmdline with the new rootfs label.

A slot designates any tegra partition, like for instance
kernel-dtb and kernel-dtb_b. This is possible currently
only for the Xavier line of devices, which do not have
u-boot support.

On Jetson Nanos there are no extra partitions on the emmc
or sd-card, only a boot bundle in the qspi flash.

On the TX2 the 32.5.1 L4T boot firmware cannot boot the newer
32.6.1 bootloaders, because the boot firmware needs to be
in sync with bootloaders and rest of the firmwares (we'll have
to try with the next L4T releases when they become available
see meta-tegra issue 879). This however is not a problem because
rollback-altboot is already accomplished in u-boot on the TX2 and
Nano devices running L4T 32.X.

Signed-off-by: Alexandru Costache <alexandru@balena.io>