tegra-binaries/tegra194-nxde-sdcard-flash: Adapt boot blob generation for NX SD 32.6.1
Starting with 32.6.1, when the device is provisioned, the
flashing tools perform bpmp trimming by default, based on
a value read from the device, which is 0. If however
the files are signed without being written, the bpmp
dtb is not trimmed and the boot blob is generated
incorrectly (wrong offsets).
Therefore we now tell the flashing tools to trim
the bpmp dtb using the default ramcode, which is also 0,
as the one that's being read from our device.

Changelog-entry: tegra-binaries/tegra194-nxde-sdcard-flash: Adapt boot blob generation for NX SD 32.6.1
Signed-off-by: Alexandru Costache <alexandru@balena.io>