conf/layer: Enable PMIC WDT for the NX Devkit SD-CARD
Cboot 32.6.1 may exhibit a sporadic hang during boot,
and the only to get out of this state is to eiter power
cycle the device manually or have the PMIC WDT enabled instead
of Tegra WDT. The sporadic hang is caused by cboot not
finding any partitions on the sd-card, and this seems
to happen right after a reboot, but very infrequently, and only
on the sd-card version of the device.

We enable the PMIC watchdog via ODMDATA on this particular
device.

Signed-off-by: Alexandru Costache <alexandru@balena.io>