cleanup: Remove community supported devices currently not maintained
We temporarily remove the following community device types
that have not been released in the dashboard because updates
have not been submitted by community maintainers in order
to keep up with newer L4T releases.

We cannot release images for these devices types without updated
L4T device-trees or configuration files that are specific to them.

These can be added back an released to the dashboard when updates
for L4T 32.6.1 or newer are submitted.

Changelog-entry: Cleanup community supported devices that are currently not maintained
Signed-off-by: Alexandru Costache <alexandru@balena.io>