cboot: Reset in case of panic instead of halting
On L4T 32.6.1 there can be a case in which cboot is
not able to load the kernel or dtb from the sd-card
and this can happen only after multiple reboots. A
manual reset is necessary for the board to boot, and
this happens only on the NX Devkit with 32.6.1.

Let's reset instead of halting in this particular case
to retry booting from the sd-card.

Signed-off-by: Alexandru Costache <alexandru@balena.io>