hostapp-update-hooks: Avoid MB1 bootloader debug mode
In the current form, rollback-altboot from
32.6.1 to 32.6.1 works on the AGX, however,
if performed from 32.5.1 to the newer one,
MB1 puts the device in debug mode instead
of loading bootloaders from the old slot.

The solution to this problem is to write
both slots with 32.6.1 and just leave the
kernel and kernel-dtb partitions on 32.5.1.

This makes MB1 successfully load the inactive
slots. For details see github link in this
commit.

Changelog-entry: Avoid MB1 bootloader debug mode
Signed-off-by: Alexandru Costache <alexandru@balena.io>