linux-tegra: Include configuration files
In commit 853224bd12f21175e273aedbc0466ac19e3aec25
we brought in the kernel recipe from meta-tegra
to fix a initramfs issue in L4T 32.7.2, until we
will have Kirkstone support in meta-balena. Currently
we use Honister for the device integration layers
and meta-tegra 32.7.2 which is compatible with Kirkstone 4.0
and Langdale 4.1.

Changelog-entry: linux-tegra: Include configuration files
Signed-off-by: Alexandru Costache <alexandru@balena.io>