jetson-dtbs: Update tegra186-tx2-6.dtb
This new version of the tegra186_tx2_6.dtb file
adds the "nvidia,rx_frames" element to the "ether_qos" block.

Alexandru: I moved the dtb added by Todd from the root
of the respository to the jetson-dtbs package, so it
can be included in the next image releases.

This is a rebase of https://github.com/balena-os/balena-jetson/pull/438

Signed-off-by: Alexandru Costache <alexandru@balena.io>
Changelog-entry: jetson-dtbs: Update tegra186-tx2-6.dtb