hostapp-update-hooks: Improve state symlinks checks
On a dfferent type of CDS devices, the state symlink to the mts-bootpack
partition may be present when hup starts but later when
dd is used to calculate the md5sum they were removed

It's possible that udev repopulates the /dev/disk/by- directories,
so let's check if the initial symlink disappeared and
perform a couple retries if it did, to see if it gets re-created or
can be obtained from the partlabel directory.
At most, if none of the symlinks exist anymore the update will fail and the
hooks from the previous OS will run. The update procedure should
be re-triggered from the API in this case.

Changelog-entry: hostapp-update-hooks: Improve state symlinks checks
Signed-off-by: Alexandru Costache <alexandru@balena.io>