hostapp-update-hooks: Improve firmware update speed
by reading/writing in larger chunks, ensure data
is flushed to disk. This reduces the risk of the
device not booting in case of an accidental power
cut during the OS update.

Also update logging to make use of the os logging
helper functions.

Signed-off-by: Alexandru Costache <alexandru@balena.io>
Changelog-entry: hostapp-update-hooks: Improve firmware update speed