workflows: Update JN30B workflow
The HUP suite will be run manually for now,
on the old hardware, because the new module
cannot be provisioned with an old image as
that image does not support the new DRAM
variant.

Also, after the v2.9.60 or newer leviathan worker release
will become available in BM, the autokit worker
will be moved there.

Signed-off-by: Alexandru Costache <alexandru@balena.io>